+++ client-app/src/admin/tests/grids/GridTestModel.ts
@@ -65,6 +65,10 @@ export class GridTestModel extends HoistModel {
     accessor includeCollapsedChildren = true;
 
     @bindable
+    @persist
+    includeHiddenColumns = false;
+
+    @bindable
     @persist.with({path: 'gridPersistType', debounce: 500}) // test persist.with!
     accessor persistType = null;
 
@@ -92,6 +96,7 @@ export class GridTestModel extends HoistModel {
                 this.autosizeMode,
                 this.renderedRowsOnly,
                 this.includeCollapsedChildren,
+                this.includeHiddenColumns,
                 this.persistType,
                 this.colChooserCommitOnChange,
                 this.colChooserShowRestoreDefaults,
@@ -206,7 +211,8 @@ export class GridTestModel extends HoistModel {
             autosizeOptions: {
                 mode: this.autosizeMode,
                 renderedRowsOnly: this.renderedRowsOnly,
-                includeCollapsedChildren: this.includeCollapsedChildren
+                includeCollapsedChildren: this.includeCollapsedChildren,
+                includeHiddenColumns: this.includeHiddenColumns
             },
             columns: [
                 {