+++ client-app/src/mobile/panels/ScrollablePanelPageModel.ts
@@ -1,6 +1,0 @@
-import {HoistModel} from '@xh/hoist/core';
-import {bindable} from '@xh/hoist/mobx';
-
-export class ScrollablePanelPageModel extends HoistModel {
-    @bindable accessor showLongContent: boolean = false;
-}